latest changes before switching to new computer

+++ AccuDrumsPlugin/Objects/GridItem.cs
@@ -9,7 +9,15 @@ namespace Accudrums.Objects {
         public byte Note { get; set; }
         public List<Sample> Samples { get; set; }
 
+        /// <summary>
+        /// The Gain value of the samples, in dB values 
+        /// </summary>
         public float Gain { get; set; } 
-        public int Panning { get; set; }
+
+        /// <summary>
+        /// Set the panning, is a value between -1 and 1
+        /// -1 is total left, 1 is total right and 0.0 is center
+        /// </summary>
+        public float Panning { get; set; }
     }
 }
